refactor(cli): share RetryingRoundTripper via pkg/kubeenv
Move implementation to pkg/kubeenv per review; kube.RetryingRoundTripper
remains a type alias for API compatibility. pkg/cli uses kubeenv only.
